Independent Components/Graphics_Engine_Main/requirements.txt
glm==0.4.3
keyboard==0.13.5
moderngl==5.7.4
numpy==1.22.4
opencv_python==4.6.0.66
Pillow==9.4.0
pygame==2.1.3
pywavefront==1.3.3
requests==2.22.0
tqdm==4.64.0
